requirements.txt
langchain==1.0.3
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-ollama==1.0.0
langchain-chroma==1.0.0

# Vector Store
chromadb==1.3.0

# Web Interface and Configuration
gradio
python-dotenv==1.0.1
pyyaml==6.0.1